try removing clear and set os envs and metadata
